s2_organisation_and_version_control/tester/DTU-02476-MLOPS-Jan-23/requirements.txt
# local package
-e .
torch>=1.8
torchvision
click
matplotlib

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
